eviction/requirements.txt
# requirements.txt pour eviction
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
